Upgraded verif initiale/finale
Changed order of buttons for verif initiale and finale and made it so
focus is switched to appropriate window when pressed.
